requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=0.19.0
atlassian-python-api>=3.41.1
openai>=1.0.0
redis>=4.5.0
tiktoken>=0.5.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development dependencies
black>=22.3.0
flake8>=4.0.1
mypy>=0.981

# API dependencies
Flask>=2.3.3
gunicorn>=21.2.0
aiohttp>=3.9.1
asyncio>=3.4.3

# Logging
python-json-logger>=2.0.7

# Type stubs - optional
types-redis

# Note: tkinter is included with Python installation
# If tkinter is missing, install Python with tkinter:
# Windows: Use Python installer and select tcl/tk option
# Linux: sudo apt-get install python3-tk
# Mac: brew install python-tk

# Core Framework
Flask==2.3.3
gunicorn==21.2.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# OpenAI Integration
openai==1.3.0

# Atlassian Integration
atlassian-python-api==3.41.2

# Caching and Session Management
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging
python-json-logger==2.0.7

# Type Checking
types-redis==4.6.0.20231124
types-requests==2.31.0.10

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4

# New dependencies for Acrolinx integration
aiohttp>=3.8.0
typing-extensions>=4.0.0

# Existing dependencies
aiohttp>=3.8.0
redis>=4.5.0
pytest-cov>=4.0.0
python-dotenv>=0.19.0
atlassian-python-api>=3.30.0 